Refactored config files to follow Single Responsibility. Ordered Guaranteed Items alphabetically.
