+++ ItemRoulette/ItemRoulette.cs
@@ -1,4 +1,5 @@
 ﻿using BepInEx;
+using ItemRoulette.Configs;
 using On.RoR2.Artifacts;
 using RoR2;
 using System.Collections.Generic;
